Make inner class a class on its own
